+++ src/components/options/index.ts
@@ -5,6 +5,5 @@ export * from './WeeklyCalendar';
 export * from './AnalyticsTab';
 export * from './PremiumTab';
 export * from './HelpTab';
-export * from './YouTubeTab';
 export * from './PasswordSettingsSection';
 export * from './modals';
